+++ src/prueba/pruebaBundle/Controller/GastonominaController.php
@@ -24,13 +24,21 @@ class GastonominaController extends Controller
      */
     public function indexAction()
     {
-        $em = $this->getDoctrine()->getManager();
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $em = $this->getDoctrine()->getManager();
 
-        $gastonominas = $em->getRepository('pruebaBundle:Gastonomina')->findAll();
+            $gastonominas = $em->getRepository('pruebaBundle:Gastonomina')->findAll();
 
-        return $this->render('gastonomina/index.html.twig', array(
-            'gastonominas' => $gastonominas,
-        ));
+            return $this->render('gastonomina/index.html.twig', array(
+                        'gastonominas' => $gastonominas,
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -41,22 +49,30 @@ class GastonominaController extends Controller
      */
     public function newAction(Request $request)
     {
-        $gastonomina = new Gastonomina();
-        $form = $this->createForm('prueba\pruebaBundle\Form\GastonominaType', $gastonomina);
-        $form->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $gastonomina = new Gastonomina();
+            $form = $this->createForm('prueba\pruebaBundle\Form\GastonominaType', $gastonomina);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($gastonomina);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($gastonomina);
+                $em->flush();
 
-            return $this->redirectToRoute('gastonomina_show', array('id' => $gastonomina->getId()));
+                return $this->redirectToRoute('gastonomina_show', array('id' => $gastonomina->getId()));
+            }
+
+            return $this->render('gastonomina/new.html.twig', array(
+                        'gastonomina' => $gastonomina,
+                        'form' => $form->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->render('gastonomina/new.html.twig', array(
-            'gastonomina' => $gastonomina,
-            'form' => $form->createView(),
-        ));
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -67,12 +83,20 @@ class GastonominaController extends Controller
      */
     public function showAction(Gastonomina $gastonomina)
     {
-        $deleteForm = $this->createDeleteForm($gastonomina);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $deleteForm = $this->createDeleteForm($gastonomina);
 
-        return $this->render('gastonomina/show.html.twig', array(
-            'gastonomina' => $gastonomina,
-            'delete_form' => $deleteForm->createView(),
-        ));
+            return $this->render('gastonomina/show.html.twig', array(
+                        'gastonomina' => $gastonomina,
+                        'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -83,23 +107,31 @@ class GastonominaController extends Controller
      */
     public function editAction(Request $request, Gastonomina $gastonomina)
     {
-        $deleteForm = $this->createDeleteForm($gastonomina);
-        $editForm = $this->createForm('prueba\pruebaBundle\Form\GastonominaType', $gastonomina);
-        $editForm->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $deleteForm = $this->createDeleteForm($gastonomina);
+            $editForm = $this->createForm('prueba\pruebaBundle\Form\GastonominaType', $gastonomina);
+            $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($gastonomina);
-            $em->flush();
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($gastonomina);
+                $em->flush();
 
-            return $this->redirectToRoute('gastonomina_edit', array('id' => $gastonomina->getId()));
+                return $this->redirectToRoute('gastonomina_edit', array('id' => $gastonomina->getId()));
+            }
+
+            return $this->render('gastonomina/edit.html.twig', array(
+                        'gastonomina' => $gastonomina,
+                        'edit_form' => $editForm->createView(),
+                        'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->render('gastonomina/edit.html.twig', array(
-            'gastonomina' => $gastonomina,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -110,16 +142,24 @@ class GastonominaController extends Controller
      */
     public function deleteAction(Request $request, Gastonomina $gastonomina)
     {
-        $form = $this->createDeleteForm($gastonomina);
-        $form->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $form = $this->createDeleteForm($gastonomina);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($gastonomina);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($gastonomina);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('gastonomina_index');
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->redirectToRoute('gastonomina_index');
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -131,10 +171,18 @@ class GastonominaController extends Controller
      */
     private function createDeleteForm(Gastonomina $gastonomina)
     {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('gastonomina_delete', array('id' => $gastonomina->getId())))
-            ->setMethod('DELETE')
-            ->getForm()
-        ;
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            return $this->createFormBuilder()
+                            ->setAction($this->generateUrl('gastonomina_delete', array('id' => $gastonomina->getId())))
+                            ->setMethod('DELETE')
+                            ->getForm()
+            ;
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 }
